requirements.txt
pandas
matplotlib
seaborn
numpy
python-box==7.0.0
ensure
PyYAML==6.0.1
Joblib==1.3.2
dataclasses
scikit-learn
-e .